Restructure project to make it usable as a framework
This restructures clin in such a way that it is easy to create custom OpenCL
interceptor libraries based on clin. The main functionality for building a
library based on clin can now be found in cmake/Findclin.cmake.

Additionally we now include a basic example that showcases how clin can be used.